% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residence.R
\name{abundance}
\alias{abundance}
\title{Create abundance within timestep}
\usage{
abundance(x, timestep = "week")
}
\arguments{
\item{x}{The object.}

\item{timestep}{A character string of the timestep. Acceptable values are those accepted by \code{lubridate::floor_date()}: second, minute, hour, day, week, month, bimonth, quarter, season, halfyear and year.}
}
\value{
A tibble.
}
\description{
Input should be output of \code{residence_event()} function with \code{squash = TRUE}.
Abundance is the total number of unque transmitters calculated from
\code{residence_proportion_complete()} grouped by receiver_group and timestep
}
