% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection-efficiency.R
\name{calculate_de}
\alias{calculate_de}
\title{Calculate detection efficiency}
\usage{
calculate_de(detection, deployment, ping_interval_s = 600, tstep = "day")
}
\arguments{
\item{detection}{A tibble of raw detection data containing columns 'transmitter_id' and 'datetime_utc'.}

\item{deployment}{A tibble of the receiver deployments containing columns 'date_deployment', 'date_last_download', 'station', 'array', 'array_rkm'.}
}
\value{
A tibble.
}
\description{
Calculates detection efficiency based on ping interval, timestep and deployments.
}
