% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.R
\name{chk_deployment}
\alias{chk_deployment}
\title{Check deployment data}
\usage{
chk_deployment(deployment)
}
\arguments{
\item{deployment}{A tibble of the receiver deployments containing columns 'date_deployment', 'date_last_download', 'station', 'array', 'array_rkm'.}
}
\value{
A flag.
}
\description{
Check deployment data
}
