% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.R
\name{chk_detection_complete}
\alias{chk_detection_complete}
\title{Check complete detection data}
\usage{
chk_detection_complete(detection_complete)
}
\arguments{
\item{detection_complete}{A tibble of complete detection data returned from \code{wsgcolr::detection_complete()} containing columns 'transmitter_id' and 'timestep'.}
}
\value{
A flag.
}
\description{
Check complete detection data
}
