% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.R
\name{chk_detection_event}
\alias{chk_detection_event}
\title{Check detection path data}
\usage{
chk_detection_event(detection_event)
}
\arguments{
\item{detection_event}{A tibble of detection path data returned from \code{wsgcolr::detection_event()}containing columns 'transmitter_id', 'timestep', 'path', 'array', 'array_rkm'.}
}
\value{
A flag.
}
\description{
Check detection path data
}
