% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.R
\name{chk_detection_path}
\alias{chk_detection_path}
\title{Check detection path data}
\usage{
chk_detection_path(detection_path)
}
\arguments{
\item{detection_path}{A tibble of detection path data containing columns 'transmitter_id', 'timestep', 'path', 'array', 'array_rkm', .}
}
\value{
A flag.
}
\description{
Check detection path data
}
