% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.R
\name{chk_detection_ratio}
\alias{chk_detection_ratio}
\title{Check detection ratio data}
\usage{
chk_detection_ratio(detection_ratio)
}
\arguments{
\item{detection_ratio}{A tibble of detection rtio data returned from \code{wsgcolr::detection_ratio()} containing columns 'timestep', 'n' and 'prop'.}
}
\value{
A flag.
}
\description{
Check detection ratio data
}
