% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.R
\name{chk_detection_timestep}
\alias{chk_detection_timestep}
\title{Check detection timestep data}
\usage{
chk_detection_timestep(detection_timestep)
}
\arguments{
\item{detection_timestep}{A tibble of aggregated detection data returned from \code{wsgcolr::detection_timestep()} containing columns 'transmitter_id' and 'timestep'. If \code{NULL}, argument is ignored.}
}
\value{
A flag.
}
\description{
Check detection timestep data
}
