% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.R
\name{chk_reference_rkm}
\alias{chk_reference_rkm}
\title{Check reference rkm locations}
\usage{
chk_reference_rkm(reference_rkm)
}
\arguments{
\item{reference_rkm}{A tibble of reference rkm locations containing columns 'label' and 'rkm'.}
}
\value{
A flag.
}
\description{
Check reference rkm locations
}
