% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.R
\name{chk_river}
\alias{chk_river}
\title{Check river data}
\usage{
chk_river(river)
}
\arguments{
\item{river}{A sf object of the river polygon. Geometry column must inherit class 'sfc_POLYGON' or 'sfc_MULTIPOLYGON'.}
}
\value{
A flag.
}
\description{
Check river data
}
