% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.R
\name{chk_timestep}
\alias{chk_timestep}
\title{Check timestep arg}
\usage{
chk_timestep(timestep)
}
\arguments{
\item{timestep}{A character string of the timestep. Acceptable values are those accepted by \code{lubridate::floor_date()}: second, minute, hour, day, week, month, bimonth, quarter, season, halfyear and year.}
}
\value{
A flag.
}
\description{
Check timestep arg
}
