% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-query-telemetry.R
\name{db_query_deployment_period}
\alias{db_query_deployment_period}
\title{Query deployment periods using dbplyr SQL}
\usage{
db_query_deployment_period(con, collect = TRUE)
}
\value{
A tibble of the queried data.
}
\description{
Get deployment start/end periods while taking into account the possibility of dead battery.
If battery is dead, gets last known detection and create new deployment period
}
