% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-query-telemetry.R
\name{db_query_deployment_period2}
\alias{db_query_deployment_period2}
\title{Query deployment periods using dbplyr SQL}
\usage{
db_query_deployment_period2(con, collect = TRUE)
}
\arguments{
\item{con}{A PqConnection object.}

\item{collect}{A flag indicating whether to collect dbplyr output into memory.}
}
\value{
A tibble of the queried data.
}
\description{
This uses much simpler logic to get deployment periods. It ignores whether
battery dead and defines the broadest possible periods just to get location of receiver
\itemize{
\item more detailed receiver coverage within each period can be obtained from receiver metadata.
}
}
