% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-query-telemetry.R
\name{db_query_detection_clean}
\alias{db_query_detection_clean}
\title{Query clean detections using dbplyr SQL}
\usage{
db_query_detection_clean(con, collect = TRUE)
}
\value{
A tibble of the queried data.
}
\description{
Converts timezone to PST and filters within deployment periods.
}
