% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-query-telemetry.R
\name{db_query_detection_timestep}
\alias{db_query_detection_timestep}
\title{Query tidy detections and summarize by timestep using dbplyr SQL}
\usage{
db_query_detection_timestep(
  con,
  timestep = "week",
  clean = TRUE,
  collect = TRUE
)
}
\arguments{
\item{con}{A PqConnection object.}

\item{timestep}{A character string of the timestep. Acceptable values are those accepted by \code{lubridate::floor_date()}: second, minute, hour, day, week, month, bimonth, quarter, season, halfyear and year.}

\item{clean}{A flag indicating whether to select the most useful columns.}

\item{collect}{A flag indicating whether to collect dbplyr output into memory.}
}
\value{
A tibble of the queried data.
}
\description{
Grouping is done by receiver_group. Receiver group with the most detections during timestep 'wins'.
}
