% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-query-environmental.R
\name{db_query_discharge_tidy}
\alias{db_query_discharge_tidy}
\title{Query tidy discharge data using dbplyr SQL
Tidy data includes joins to discharge_station to get flow_zone}
\usage{
db_query_discharge_tidy(con, clean = TRUE, collect = TRUE)
}
\arguments{
\item{con}{A PqConnection object.}

\item{clean}{A flag indicating whether to select the most useful columns.}

\item{collect}{A flag indicating whether to collect dbplyr output into memory.}
}
\value{
A tibble of the queried data.
}
\description{
Query tidy discharge data using dbplyr SQL
Tidy data includes joins to discharge_station to get flow_zone
}
