% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-read.R
\name{db_read}
\alias{db_read}
\title{Read table from database connection}
\usage{
db_read(con, table, sf = FALSE, collect = TRUE)
}
\value{
A data.frame of the data.
}
\description{
Read table from database connection
}
