% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-read.R
\name{db_read_detection_daily}
\alias{db_read_detection_daily}
\title{Read detection_tidy_daily View}
\usage{
db_read_detection_daily(conn)
}
\value{
A tibble of the summarized detection data.
}
\description{
Read detection_tidy_daily View
}
