% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-read.R
\name{db_read_detection_tidy}
\alias{db_read_detection_tidy}
\title{Read detection_tidy_weekly View}
\usage{
db_read_detection_tidy(con, timestep = "week", transmitter_id = NULL)
}
\arguments{
\item{timestep}{A character string of the timestep. Acceptable values are those accepted by \code{lubridate::floor_date()}: second, minute, hour, day, week, month, bimonth, quarter, season, halfyear and year.}
}
\value{
A tibble of the summarized detection data.
}
\description{
Read detection_tidy_weekly View
}
