% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-read.R
\name{db_read_receiver_group}
\alias{db_read_receiver_group}
\title{Read receiver_group table and add receiver_group_colour}
\usage{
db_read_receiver_group(con)
}
\value{
A tibble of the summarized detection data.
}
\description{
Read receiver_group table and add receiver_group_colour
}
