% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-read.R
\name{db_read_station}
\alias{db_read_station}
\title{Read detection_tidy_weekly View}
\usage{
db_read_station(con, sf = TRUE)
}
\arguments{
\item{con}{A PqConnection object.}

\item{sf}{A flag indicating whether to provide output as sf object.}
}
\value{
A tibble of the summarized detection data.
}
\description{
Read detection_tidy_weekly View
}
