% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-write.R
\name{db_write}
\alias{db_write}
\title{Write data to database connection}
\usage{
db_write(con, table, data, append = TRUE, ...)
}
\arguments{
\item{con}{A PqConnection object.}

\item{table}{A string of the schema and table name separated by a '.' (e.g., "telemetry.detection")}

\item{data}{A data.frame of the data to write to database.}

\item{append}{A flag indicating whether to append data to database table.}

\item{...}{Arguments passed to another function.}
}
\value{
The modified database.
}
\description{
Write data to database connection
}
