% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-write-telemetry.R
\name{db_write_deployment}
\alias{db_write_deployment}
\title{Write deployment csv to db}
\usage{
db_write_deployment(file, con)
}
\value{
The modified database
}
\description{
Write deployment csv to db
}
