% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-write-environment.R
\name{db_write_discharge_poisson}
\alias{db_write_discharge_poisson}
\title{Write discharge csv queried from Poisson db}
\usage{
db_write_discharge_poisson(con, file, tz_data = "Etc/GMT+8")
}
\value{
The modified database
}
\description{
Write discharge csv queried from Poisson db
}
