% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-write-telemetry.R
\name{db_write_temperature_deployment}
\alias{db_write_temperature_deployment}
\title{Write temperature deployment to db}
\usage{
db_write_temperature_deployment(con, x)
}
\arguments{
\item{con}{A PqConnection object.}

\item{x}{The object.}
}
\value{
The modified database
}
\description{
Write temperature deployment to db
}
