% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-write-telemetry.R
\name{db_write_transmitter}
\alias{db_write_transmitter}
\title{Write transmitter csv file to db}
\usage{
db_write_transmitter(file, con)
}
\value{
The modified database
}
\description{
Write transmitter csv file to db
}
