% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection-ratio.R
\name{detection_complete}
\alias{detection_complete}
\title{Get complete detection data (i.e. filling in absent receiver_group/timestep)}
\usage{
detection_complete(
  detection_timestep,
  timestep = "week",
  receiver_group = "receiver_group"
)
}
\arguments{
\item{detection_timestep}{A tibble of aggregated detection data returned from \code{wsgcolr::detection_timestep()} containing columns 'transmitter_id' and 'timestep'. If \code{NULL}, argument is ignored.}

\item{timestep}{A character string of the timestep. Acceptable values are those accepted by \code{lubridate::floor_date()}: second, minute, hour, day, week, month, bimonth, quarter, season, halfyear and year.}

\item{receiver_group}{A character string of the column containing receiver group (e.g., station, zone, array).}
}
\value{
A modified tibble of the complete detection data.
}
\description{
Get complete detection data (i.e. filling in absent receiver_group/timestep)
}
