% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate-detection.R
\name{detection_event}
\alias{detection_event}
\title{Assign detections to events and paths}
\usage{
detection_event(
  detection_timestep,
  receiver_group = "array",
  max_absence = 96,
  squash = FALSE
)
}
\arguments{
\item{detection_timestep}{A tibble of aggregated detection data returned from \code{wsgcolr::detection_timestep()} containing columns 'transmitter_id' and 'timestep'. If \code{NULL}, argument is ignored.}

\item{receiver_group}{A character string of the column containing receiver group (e.g., station, zone, array).}

\item{max_absence}{A number of the maximum hours without detection for new event to occur.}

\item{squash}{A flag indicating whether to reduce detection events to first and last timesteps.}
}
\value{
A modified tibble of the detection events and paths.
}
\description{
A new detection event occurs if fish moves to new receiver group or if more than \code{max_absence} time passes without a detection.
A new detection path occurs if more than \code{max_absence} time passes without a detection (i.e., regardless of change in receiver group)
}
