% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate-detection.R
\name{detection_ratio}
\alias{detection_ratio}
\title{Detection ratio data}
\usage{
detection_ratio(detection_complete, receiver_group = "array")
}
\arguments{
\item{detection_complete}{A tibble of complete detection data returned from \code{wsgcolr::detection_complete()} containing columns 'transmitter_id' and 'timestep'.}

\item{receiver_group}{A character string of the column containing receiver group (e.g., station, zone, array).}
}
\value{
A modified tibble of the detection ratios.
}
\description{
Create absolute (i.e., number of fish) and relative (i.e., proportion of fish) detection ratios per timestep/receiver group.
}
