% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detection-ratio.R
\name{detection_residence}
\alias{detection_residence}
\title{Calculate detection residence.}
\usage{
detection_residence(
  detection_event,
  timestep = "week",
  receiver_group = "receiver_group"
)
}
\arguments{
\item{detection_event}{A tibble of detection path data returned from \code{wsgcolr::detection_event()}containing columns 'transmitter_id', 'timestep', 'path', 'array', 'array_rkm'.}

\item{timestep}{A character string of the timestep. Acceptable values are those accepted by \code{lubridate::floor_date()}: second, minute, hour, day, week, month, bimonth, quarter, season, halfyear and year.}

\item{receiver_group}{A character string of the column containing receiver group (e.g., station, zone, array).}
}
\value{
A tibble.
}
\description{
Determine detection presence/absence for each transmitter/timestep/receiver group combination
\code{timestep} determines how to sequence date(times) between event start and event end - ideally this should match what was used in \code{detection_timestep()}.
}
