% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate-detection.R
\name{detection_timestep}
\alias{detection_timestep}
\title{Aggregate detection by timestep}
\usage{
detection_timestep(detection, timestep = "day", receiver_group = "array")
}
\arguments{
\item{detection}{A tibble of raw detection data containing columns 'transmitter_id' and 'datetime_utc'.}

\item{timestep}{A character string of the timestep. Acceptable values are those accepted by \code{lubridate::floor_date()}: second, minute, hour, day, week, month, bimonth, quarter, season, halfyear and year.}

\item{receiver_group}{A character string of the column containing receiver group (e.g., station, zone, array).}
}
\value{
A modified tibble of the detection data.
}
\description{
Receiver group (location_column) per transmitter/timestep assigned by greatest number of detections.
}
