% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-residence.R
\name{plot_abundance}
\alias{plot_abundance}
\title{Plot abundance as heatmap}
\usage{
plot_abundance(x, proportion = TRUE, size = 0.3)
}
\arguments{
\item{x}{The object.}
}
\value{
An object of class 'ggplot'.
}
\description{
Input data should be output of \code{abundance()} function. Abundance
calculations are derived from residence events and specified timestep.
}
