% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-abundance.R
\name{plot_abundance_proportion}
\alias{plot_abundance_proportion}
\title{Plot abundance heatmap as proportion of total tagged fish.}
\usage{
plot_abundance_proportion(
  x,
  xlab = "Timestep",
  ylab = "Receiver Group",
  legend_lab = "Abundance\\nProportion",
  size = 0.3
)
}
\arguments{
\item{x}{The object.}
}
\value{
An object of class 'ggplot'.
}
\description{
Input data should be output of \code{abundance()} function. Abundance
calculations are derived from residence events and specified timestep.
}
\examples{
\dontrun{
event <- residence_event(detection)
x <- abundance(event)
plot_abundance_proportion(x)
}
}
