% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-deployment.R
\name{plot_deployment}
\alias{plot_deployment}
\title{Plot deployment}
\usage{
plot_deployment(deployment, detection = NULL)
}
\arguments{
\item{deployment}{A tibble of the receiver deployments containing columns 'date_deployment', 'date_last_download', 'station', 'array', 'array_rkm'.}

\item{detection}{A tibble of detection data containing columns 'transmitter_id', 'timestep', 'station'. If \code{NULL}, argument is ignored.}
}
\value{
An object of class 'ggplot'.
}
\description{
Plot receiver deployments over time.
Optionally provide detection data to visualize detections overlapping deployment periods.
}
\examples{
\dontrun{
plot_deployment(deployment)
}
}
