% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-deployment.R
\name{plot_deployment_spatial}
\alias{plot_deployment_spatial}
\title{Plot station map}
\usage{
plot_deployment_spatial(station, river)
}
\arguments{
\item{station}{A sf object of the station point locations containing column 'array' (for color coding). Geometry column must inherit class 'sfc_POINT' or 'sfc_MULTIPOINT'.}

\item{river}{A sf object of the river polygon. Geometry column must inherit class 'sfc_POLYGON' or 'sfc_MULTIPOLYGON'.}
}
\value{
An object of class 'ggplot'.
}
\description{
Plot receiver station locations in Columbia and Kootenay River.
}
\examples{
\dontrun{
plot_deployment_spatial(station, river)
}
}
