% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-detection.R
\name{plot_detection_path}
\alias{plot_detection_path}
\title{Plot detection paths}
\usage{
plot_detection_path(
  detection_path,
  deployment,
  reference_locations,
  lims_x,
  lims_y
)
}
\arguments{
\item{detection_path}{A tibble of detection path data containing columns 'transmitter_id', 'timestep', 'path', 'array', 'array_rkm', .}

\item{deployment}{A tibble of the receiver deployments containing columns 'date_deployment', 'date_last_download', 'station', 'array', 'array_rkm'.}

\item{reference_locations}{A tibble of reference locations containing columns 'label' and 'rkm'.}

\item{lims_x}{A vector of class 'Date' containing x-axis limits.}

\item{lims_y}{A vector of the numeric y-axis limits.}
}
\value{
An object of class 'ggplot'.
}
\description{
Plot Columbia River White Sturgeon movement paths from VR2 receiver detection data.
}
\examples{
\dontrun{
lims_x = c(min(deployment$date_deployment), max(deployment$date_last_download))
lims_y = c(0, 56)
plot_detection_path(detection, deployment, reference_locations)#'
}
}
