% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-detection-ratio.R
\name{plot_detection_ratio_absolute}
\alias{plot_detection_ratio_absolute}
\title{Plot absolute detection ratio}
\usage{
plot_detection_ratio_absolute(detection_ratio, receiver_group = "array")
}
\arguments{
\item{detection_ratio}{A tibble of detection rtio data returned from \code{wsgcolr::detection_ratio()} containing columns 'timestep', 'n' and 'prop'.}

\item{receiver_group}{A character string of the column containing receiver group (e.g., station, zone, array).}
}
\value{
An object of class 'ggplot'.
}
\description{
Plot stacked area chart for absolute detection ratios (i.e., number of fish per timestep/receiver group).
}
\examples{
\dontrun{
plot_detection_ratio_absolute(detection_ratio)
}
}
