% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-discharge.R
\name{plot_discharge}
\alias{plot_discharge}
\title{Plot Deployment}
\usage{
plot_discharge(x, by = "station_id", ...)
}
\arguments{
\item{x}{The object.}

\item{by}{A string of the column name to group by.}

\item{...}{Arguments passed to another function.}
}
\description{
Plot discharge by station.
}
