% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-residence.R
\name{plot_residence_event}
\alias{plot_residence_event}
\title{Plot residence event}
\usage{
plot_residence_event(
  x,
  residence_path = NULL,
  deployment = NULL,
  reference_rkm = reference_rkms(),
  datetime = "datetime_pst",
  receiver_group = "receiver_group",
  receiver_group_rkm = "receiver_group_rkm",
  xlims = range(x$event_start),
  ylims = levels(x$mean_rkm),
  ...
)
}
\arguments{
\item{x}{The object.}

\item{deployment}{A tibble of the receiver deployments containing columns 'date_deployment', 'date_last_download', 'station', 'array', 'array_rkm'.}

\item{reference_rkm}{A tibble of reference rkm locations containing columns 'label' and 'rkm'.}

\item{datetime}{A string of the column name containing datetimes.}

\item{receiver_group}{A character string of the column containing receiver group (e.g., station, zone, array).}

\item{receiver_group_rkm}{A character string of the column containing receiver group rkm.}

\item{...}{Arguments passed to another function.}
}
\value{
An object of class 'ggplot'.
}
\description{
Plot resident events of a single transmitter.
Optionally provide resident_path as geom_line and deployments as grey geom_segments.
}
\examples{
\dontrun{
x <- residence_event(detection)
plot_residence_event(x)
}
}
