% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-environmental.R
\name{read_file_detection}
\alias{read_file_detection}
\title{read raw VUE detection csv export and prepare for db}
\usage{
read_file_detection(file, clean = TRUE)
}
\arguments{
\item{file}{A string of the file path.}

\item{clean}{A flag indicating whether to select the most useful columns.}
}
\value{
A tibble
}
\description{
read raw VUE detection csv export and prepare for db
}
