% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-environmental.R
\name{read_file_discharge_poisson}
\alias{read_file_discharge_poisson}
\title{read discharge csv queried from Poisson db and prepare for wsgcolr db}
\usage{
read_file_discharge_poisson(file, tz_data = "Etc/GMT+8")
}
\arguments{
\item{file}{A string of the file path.}

\item{tz_data}{A string of the timezone of the data.}
}
\value{
A tibble
}
\description{
read discharge csv queried from Poisson db and prepare for wsgcolr db
}
