% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-environmental.R
\name{read_file_temperature}
\alias{read_file_temperature}
\title{Read weird temperature file and extract logger and station}
\usage{
read_file_temperature(x, tz_data = "Etc/GMT+8")
}
\arguments{
\item{x}{The object.}

\item{tz_data}{A string of the timezone of the data.}
}
\value{
A tibble.
}
\description{
Read weird temperature file and extract logger and station
}
