% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-environmental.R
\name{read_files}
\alias{read_files}
\title{Read files using reader function and bind together}
\usage{
read_files(x, fun = read_temperature_file, quiet = FALSE)
}
\arguments{
\item{x}{The object.}

\item{quiet}{A flag indicating whether to suppress warnings and messages.}
}
\value{
A tibble.
}
\description{
Read files using reader function and bind together
}
