% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-temperature.R
\name{read_temperature_file}
\alias{read_temperature_file}
\title{Read weird temperature file and extract logger and station}
\usage{
read_temperature_file(x, tz = "PST8PDT")
}
\arguments{
\item{x}{The object.}

\item{tz}{A string of the timezone.}
}
\value{
A tibble.
}
\description{
Read weird temperature file and extract logger and station
}
