% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residence.R
\name{residence_complete}
\alias{residence_complete}
\title{Complete residence proportion}
\usage{
residence_complete(x, timestep)
}
\arguments{
\item{x}{The object.}

\item{timestep}{A character string of the timestep. Acceptable values are those accepted by \code{lubridate::floor_date()}: second, minute, hour, day, week, month, bimonth, quarter, season, halfyear and year.}
}
\value{
A tibble.
}
\description{
Input should be output of \code{residence_proportion()}.
Possible absent timesteps are determined from first and last detection for each
transmitter. Possible absent receiver_groups are determined from factor levels of receiver_group column.
}
