% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residence.R
\name{residence_event}
\alias{residence_event}
\title{Create detection residence events}
\usage{
residence_event(
  detection,
  max_absence = 96,
  min_detections = 2,
  min_duration = 6,
  squash = TRUE,
  receiver_group = "receiver_group",
  datetime = "datetime_pst"
)
}
\arguments{
\item{detection}{A tibble of raw detection data containing columns 'transmitter_id' and 'datetime_utc'.}

\item{max_absence}{A number of the maximum hours without detection for new event to occur.}

\item{squash}{A flag indicating whether to reduce detection events to first and last timesteps.}

\item{receiver_group}{A character string of the column containing receiver group (e.g., station, zone, array).}

\item{datetime}{A string of the column name containing datetimes.}
}
\value{
A tibble.
}
\description{
A new residence event occurs if a fish moves to a new receiver group, more than
\code{max_absence} time passes without a detection, more than \code{min_detections} occurs
and more than \code{min_duration} occurs.
}
