% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residence.R
\name{residence_proportion}
\alias{residence_proportion}
\title{Create residence proportion within timesteps}
\usage{
residence_proportion(x, timestep = "week", combine_transmitters = FALSE)
}
\arguments{
\item{x}{The object.}

\item{timestep}{A character string of the timestep. Acceptable values are those accepted by \code{lubridate::floor_date()}: second, minute, hour, day, week, month, bimonth, quarter, season, halfyear and year.}
}
\value{
A tibble.
}
\description{
Input should be output of \code{residence_event()} function with \code{squash = TRUE}.
Residence proportion is calculated for each transmitter, receiver_group,
and timestep where fish is present (i.e., from residence events).
If \code{combine_transmitters = TRUE}, mean residence proportion is calculated for
all transmitters.
}
