% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chk.R
\name{chk_detection}
\alias{chk_detection}
\title{Check detection data}
\usage{
chk_detection(detection)
}
\arguments{
\item{detection}{A tibble of raw detection data containing columns 'transmitter_id' and 'datetime_utc'.}
}
\value{
A flag.
}
\description{
Check detection data
}
