% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-query-telemetry.R
\name{db_query_capture_tidy}
\alias{db_query_capture_tidy}
\title{Query tidy captures using dbplyr SQL}
\usage{
db_query_capture_tidy(con, collect = TRUE)
}
\arguments{
\item{con}{A PqConnection object.}

\item{collect}{A flag indicating whether to collect dbplyr output into memory.}
}
\value{
A tibble of the queried data.
}
\description{
Joins capture table to transmitter and fish.
}
