% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-query-telemetry.R
\name{db_query_detection_station}
\alias{db_query_detection_station}
\title{Query clean detections and summarize by station and timestep}
\usage{
db_query_detection_station(con, time_step = "week", collect = TRUE)
}
\value{
A tibble of queried data.
}
\description{
This is useful for visualizing deployments as it loses information about tag.
}
