% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-query-telemetry.R
\name{db_query_detection_tidy}
\alias{db_query_detection_tidy}
\title{Query tidy detections using dbplyr SQL}
\usage{
db_query_detection_tidy(con, clean = TRUE, collect = TRUE)
}
\arguments{
\item{con}{A PqConnection object.}

\item{clean}{A flag indicating whether to select the most useful columns.}

\item{collect}{A flag indicating whether to collect dbplyr output into memory.}
}
\value{
A tibble of the queried data.
}
\description{
Joins clean detections to tidy captures, station and receiver_group.
}
