% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-query-telemetry.R
\name{db_query_detection_timestep}
\alias{db_query_detection_timestep}
\title{Query tidy detections and summarize by timestep using dbplyr SQL}
\usage{
db_query_detection_timestep(
  con,
  time_step = "week",
  clean = TRUE,
  collect = TRUE
)
}
\value{
A tibble of the queried data.
}
\description{
Grouping is done by receiver_group. Receiver group with the most detections during timestep 'wins'.
}
