% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-query-environmental.R
\name{db_query_temperature_clean}
\alias{db_query_temperature_clean}
\title{Query clean temperature data using dbplyr SQL
Clean data is filtered within deployments}
\usage{
db_query_temperature_clean(con, collect = TRUE)
}
\value{
A tibble of the queried data.
}
\description{
Query clean temperature data using dbplyr SQL
Clean data is filtered within deployments
}
