% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-write-telemetry.R
\name{db_write_detection}
\alias{db_write_detection}
\title{Write raw VUE detection csv export to db}
\usage{
db_write_detection(con, file, clean = TRUE)
}
\arguments{
\item{con}{A PqConnection object.}

\item{file}{A string of the file path.}

\item{clean}{A flag indicating whether to select the most useful columns.}
}
\value{
The modified database
}
\description{
Write raw VUE detection csv export to db
}
