% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db-write-environment.R
\name{db_write_discharge_border}
\alias{db_write_discharge_border}
\title{Write discharge csv from border flows}
\usage{
db_write_discharge_border(con, file, tz_data = "Etc/GMT+8")
}
\value{
The modified database
}
\description{
Write discharge csv from border flows
}
