% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manipulate-detection.R
\name{detection_complete}
\alias{detection_complete}
\title{Complete detection data}
\usage{
detection_complete(
  detection_timestep,
  timestep = "day",
  receiver_group = "array"
)
}
\arguments{
\item{detection_timestep}{A tibble of aggregated detection data returned from \code{wsgcolr::detection_timestep()} containing columns 'transmitter_id' and 'timestep'. If \code{NULL}, argument is ignored.}

\item{timestep}{A character string of the timestep. Acceptable values are those accepted by \code{lubridate::floor_date()}: second, minute, hour, day, week, month, bimonth, quarter, season, halfyear and year.}

\item{receiver_group}{A character string of the column containing receiver group (e.g., station, zone, array).}
}
\value{
A modified tibble of the complete detection data.
}
\description{
Create detection presence/absence for each transmitter/timestep/receiver group combination
\code{timestep} should match what was used in \code{detection_timestep}.
}
