% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/params.R
\name{params}
\alias{params}
\title{Parameter Descriptions for wsgcolr functions}
\arguments{
\item{detection_event}{A tibble of detection path data returned from \code{wsgcolr::detection_event()}containing columns 'transmitter_id', 'timestep', 'path', 'array', 'array_rkm'.}

\item{detection_timestep}{A tibble of aggregated detection data returned from \code{wsgcolr::detection_timestep()} containing columns 'transmitter_id' and 'timestep'. If \code{NULL}, argument is ignored.}

\item{detection_complete}{A tibble of complete detection data returned from \code{wsgcolr::detection_complete()} containing columns 'transmitter_id' and 'timestep'.}

\item{detection_ratio}{A tibble of detection rtio data returned from \code{wsgcolr::detection_ratio()} containing columns 'timestep', 'n' and 'prop'.}

\item{detection}{A tibble of raw detection data containing columns 'transmitter_id' and 'datetime_utc'.}

\item{receiver_group}{A character string of the column containing receiver group (e.g., station, zone, array).}

\item{receiver_group_rkm}{A character string of the column containing receiver group rkm.}

\item{reference_rkm}{A tibble of reference rkm locations containing columns 'label' and 'rkm'.}

\item{timestep}{A character string of the timestep. Acceptable values are those accepted by \code{lubridate::floor_date()}: second, minute, hour, day, week, month, bimonth, quarter, season, halfyear and year.}

\item{max_absence}{A number of the maximum hours without detection for new event to occur.}

\item{squash}{A flag indicating whether to reduce detection events to first and last timesteps.}

\item{deployment}{A tibble of the receiver deployments containing columns 'date_deployment', 'date_last_download', 'station', 'array', 'array_rkm'.}

\item{station}{A sf object of the station point locations containing column 'array' (for color coding). Geometry column must inherit class 'sfc_POINT' or 'sfc_MULTIPOINT'.}

\item{river}{A sf object of the river polygon. Geometry column must inherit class 'sfc_POLYGON' or 'sfc_MULTIPOLYGON'.}

\item{con}{A PqConnection object.}

\item{x}{The object.}

\item{pool}{A flag indicating whether to pool databse connection.}

\item{collect}{A flag indicating whether to collect dbplyr output into memory.}

\item{clean}{A flag indicating whether to select the most useful columns.}

\item{sf}{A flag indicating whether to provide output as sf object.}

\item{table}{A string of the schema and table name separated by a '.' (e.g., "telemetry.detection")}

\item{append}{A flag indicating whether to append data to database table.}

\item{data}{A data.frame of the data to write to database.}

\item{file}{A string of the file path.}

\item{tz_data}{A string of the timezone of the data.}

\item{station_col}{A string of the column name with station names.}

\item{transmitter_id}{A character vector of transmitter_id.}

\item{tz}{A string of the timezone.}

\item{quiet}{A flag indicating whether to suppress warnings and messages.}

\item{dbname}{A string of the database name.}

\item{schema}{A string of the schema name.}

\item{...}{Arguments passed to another function.}

\item{by}{A string of the column name to group by.}

\item{datetime}{A string of the column name containing datetimes.}

\item{ignore_movement}{A flag indicating whether to ignore movement between receiver groups when calculating detection events.}
}
\description{
Parameter Descriptions for wsgcolr functions
}
\keyword{internal}
