% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-deployment.R
\name{plot_deployment}
\alias{plot_deployment}
\alias{plot_deployment.data.frame}
\alias{plot_deployment.PqConnection}
\title{Plot Deployment}
\usage{
plot_deployment(x, ...)

\method{plot_deployment}{data.frame}(
  x,
  station,
  river,
  detection = NULL,
  station_col = "station_name",
  ...
)

\method{plot_deployment}{PqConnection}(x, detection = FALSE, ...)
}
\arguments{
\item{x}{The object.}

\item{...}{Arguments passed to another function.}

\item{station}{A sf object of the station point locations containing column 'array' (for color coding). Geometry column must inherit class 'sfc_POINT' or 'sfc_MULTIPOINT'.}

\item{river}{A sf object of the river polygon. Geometry column must inherit class 'sfc_POLYGON' or 'sfc_MULTIPOLYGON'.}

\item{detection}{A tibble of raw detection data containing columns 'transmitter_id' and 'datetime_utc'.}

\item{station_col}{A string of the column name with station names.}
}
\description{
Generic function to plot the deployment periods.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: method to plot deployment periods from data.frames

\item \code{PqConnection}: method to plot deployments fron connection object
}}

\examples{

\dontrun{
plot_deployment(deployment, station, river)
}

\dontrun{
plot_deployment(con)
}
}
