% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-detection.R
\name{plot_detection_path}
\alias{plot_detection_path}
\alias{plot_detection_path.data.frame}
\alias{plot_detection_path.PqConnection}
\title{Plot Detection Path}
\usage{
plot_detection_path(x, ...)

\method{plot_detection_path}{data.frame}(
  detection_event,
  deployment,
  reference_rkm = reference_rkms(),
  receiver_group = "receiver_group",
  receiver_group_rkm = "receiver_group_rkm",
  lims_x = range(detection_event$timestep),
  lims_y = range(detection_event$receiver_group_rkm)
)

\method{plot_detection_path}{PqConnection}(
  con,
  transmitter_id,
  max_absence = 96,
  lims_x = c(as.Date("2010-01-01"), as.Date("2020-01-01")),
  lims_y = c(0, 56)
)
}
\arguments{
\item{detection_event}{A tibble of detection path data returned from \code{wsgcolr::detection_event()}containing columns 'transmitter_id', 'timestep', 'path', 'array', 'array_rkm'.}

\item{deployment}{A tibble of the receiver deployments containing columns 'date_deployment', 'date_last_download', 'station', 'array', 'array_rkm'.}

\item{reference_rkm}{A tibble of reference rkm locations containing columns 'label' and 'rkm'.}

\item{receiver_group}{A character string of the column containing receiver group (e.g., station, zone, array).}

\item{receiver_group_rkm}{A character string of the column containing receiver group rkm.}

\item{lims_x}{A vector of class 'Date' containing x-axis limits.}

\item{lims_y}{A vector of the numeric y-axis limits.}

\item{max_absence}{A number of the maximum hours without detection for new event to occur.}
}
\value{
An object of class 'ggplot'.

An object of class 'ggplot'.
}
\description{
Generic function to plot the detection paths.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: method to plot detection paths from data.frames

\item \code{PqConnection}: method to plot detection paths from data.frames
}}

\examples{
\dontrun{
lims_x = c(min(deployment$date_deployment), max(deployment$date_last_download))
lims_y = c(0, 56)
plot_detection_path(detection, deployment, reference_rkm)
}
\dontrun{
lims_x = c(min(deployment$date_deployment), max(deployment$date_last_download))
lims_y = c(0, 56)
plot_detection_path(con, transmitter_id = "A69-1303-10060", lims_x = lims_x, lims_y = lims_y)
}
}
