% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-discharge.R
\name{plot_discharge}
\alias{plot_discharge}
\alias{plot_discharge.data.frame}
\alias{plot_discharge.PqConnection}
\title{Plot Deployment}
\usage{
plot_discharge(x, ...)

\method{plot_discharge}{data.frame}(x, by = "station_id")

\method{plot_discharge}{PqConnection}(
  con,
  station = c("HLK_ALH", "BRD_BRDS_BRX", "US_CAN", "BIR_QR"),
  by = "station_id"
)
}
\arguments{
\item{station}{A sf object of the station point locations containing column 'array' (for color coding). Geometry column must inherit class 'sfc_POINT' or 'sfc_MULTIPOINT'.}
}
\description{
Generic function to plot discharge by station.
}
\section{Methods (by class)}{
\itemize{
\item \code{data.frame}: method to plot discharge from data.frame

\item \code{PqConnection}: method to plot discharge from db connection
}}

