% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-residence.R
\name{plot_residence_proportion}
\alias{plot_residence_proportion}
\title{Plot residence proportion as heatmap}
\usage{
plot_residence_proportion(x, size = 0.3)
}
\arguments{
\item{x}{The object.}
}
\value{
An object of class 'ggplot'.
}
\description{
Input data should be output of \code{residence_proportion()} function.
Residence proportion calculations are derived from residence events
and specified timestep.
}
