% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read-temperature.R
\name{read_temperature_files}
\alias{read_temperature_files}
\title{Read weird temperature files and extract logger and station}
\usage{
read_temperature_files(x, quiet = FALSE)
}
\arguments{
\item{x}{The object.}

\item{quiet}{A flag indicating whether to suppress warnings and messages.}
}
\value{
A tibble.
}
\description{
Read weird temperature files and extract logger and station
}
