% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot-deployment.R
\name{rkm_colour}
\alias{rkm_colour}
\title{Add viridis palette to receiver group by rkm}
\usage{
rkm_colour(x)
}
\arguments{
\item{x}{The object.}
}
\value{
A vector of colours
}
\description{
Add viridis palette to receiver group by rkm
}
